Fjernet siste rest av redux
