--- src/store/rootState.ts
@@ -1,11 +1,0 @@
-import { combineReducers, Reducer } from 'redux';
-import { HelseSpionState } from '../data/types/sporenstreksTypes';
-import { helseSpionReducer } from './reducers/helseSpionReducers';
-
-export interface RootState {
-  helseSpionState: HelseSpionState,
-}
-
-export const rootReducers: Reducer<RootState> = combineReducers({
-  helseSpionState: helseSpionReducer,
-});
